backend/requirements.txt
alembic==1.13.1
APScheduler==3.10.4
Flask==3.0.3
Flask_Cors==4.0.0
Flask_Migrate==4.0.5
flask_session==0.8.0
flask_sqlalchemy==3.1.1
mysqlclient==2.2.4
pydash==8.0.0
python_dateutil==2.9.0.post0
Requests==2.31.0
SQLAlchemy==2.0.27
